+++ kernel/bio.c
@@ -18,7 +18,7 @@
 #include "param.h"
 #include "spinlock.h"
 #include "sleeplock.h"
-#include "riscv.h"
+#include "arch/riscv.h"
 #include "defs.h"
 #include "fs.h"
 #include "buf.h"
