+++ kernel/console.c
@@ -18,7 +18,7 @@
 #include "fs.h"
 #include "file.h"
 #include "memlayout.h"
-#include "riscv.h"
+#include "arch/riscv.h"
 #include "defs.h"
 #include "proc.h"
