+++ kernel/dev/virtio_disk.c
@@ -7,7 +7,7 @@
 //
 
 #include "types.h"
-#include "riscv.h"
+#include "arch/riscv.h"
 #include "defs.h"
 #include "param.h"
 #include "memlayout.h"
