+++ kernel/fs/file.c
@@ -3,7 +3,7 @@
 //
 
 #include "types.h"
-#include "riscv.h"
+#include "arch/riscv.h"
 #include "defs.h"
 #include "param.h"
 #include "fs.h"
