+++ kernel/include/net/arp.h
@@ -1,0 +1,21 @@
+// an ARP packet (comes after an Ethernet header).
+struct arp {
+  uint16 hrd; // format of hardware address
+  uint16 pro; // format of protocol address
+  uint8  hln; // length of hardware address
+  uint8  pln; // length of protocol address
+  uint16 op;  // operation
+
+  char   sha[ETHADDR_LEN]; // sender hardware address
+  uint32 sip;              // sender IP address
+  char   tha[ETHADDR_LEN]; // target hardware address
+  uint32 tip;              // target IP address
+} __attribute__((packed));
+
+#define ARP_HRD_ETHER 1 // Ethernet
+
+enum {
+  ARP_OP_REQUEST = 1, // requests hw addr given protocol addr
+  ARP_OP_REPLY = 2,   // replies a hw addr given protocol addr
+};
+
