+++ kernel/include/net/dns.h
@@ -1,0 +1,35 @@
+// an DNS packet (comes after an UDP header).
+struct dns {
+  uint16 id;  // request ID
+
+  uint8 rd: 1;  // recursion desired
+  uint8 tc: 1;  // truncated
+  uint8 aa: 1;  // authoritive
+  uint8 opcode: 4; 
+  uint8 qr: 1;  // query/response
+  uint8 rcode: 4; // response code
+  uint8 cd: 1;  // checking disabled
+  uint8 ad: 1;  // authenticated data
+  uint8 z:  1;  
+  uint8 ra: 1;  // recursion available
+  
+  uint16 qdcount; // number of question entries
+  uint16 ancount; // number of resource records in answer section
+  uint16 nscount; // number of NS resource records in authority section
+  uint16 arcount; // number of resource records in additional records
+} __attribute__((packed));
+
+struct dns_question {
+  uint16 qtype;
+  uint16 qclass;
+} __attribute__((packed));
+  
+#define ARECORD (0x0001)
+#define QCLASS  (0x0001)
+
+struct dns_data {
+  uint16 type;
+  uint16 class;
+  uint32 ttl;
+  uint16 len;
+} __attribute__((packed));
