+++ kernel/include/net/ethernet.h
@@ -1,0 +1,13 @@
+
+#define ETHADDR_LEN 6
+
+// an Ethernet packet header (start of the packet).
+struct eth {
+  uint8  dhost[ETHADDR_LEN];
+  uint8  shost[ETHADDR_LEN];
+  uint16 type;
+} __attribute__((packed));
+
+#define ETHTYPE_IP  0x0800 // Internet protocol
+#define ETHTYPE_ARP 0x0806 // Address resolution protocol
+
