+++ kernel/include/net/udp.h
@@ -1,0 +1,9 @@
+// a UDP packet header (comes after an IP header).
+struct udp {
+  uint16 sport; // source port
+  uint16 dport; // destination port
+  uint16 ulen;  // length, including udp header, not including IP header
+  uint16 sum;   // checksum
+};
+
+
