+++ kernel/lib/kalloc.c
@@ -6,7 +6,7 @@
 #include "param.h"
 #include "memlayout.h"
 #include "spinlock.h"
-#include "riscv.h"
+#include "arch/riscv.h"
 #include "defs.h"
 
 void freerange(void *pa_start, void *pa_end);
