+++ kernel/lib/string.c
