+++ kernel/net/Makefile
@@ -1,0 +1,5 @@
+KROOT=../
+
+NETOBJS = \
+	
+net.o:
