+++ kernel/proc/pipe.c
@@ -1,5 +1,5 @@
 #include "types.h"
-#include "riscv.h"
+#include "arch/riscv.h"
 #include "defs.h"
 #include "param.h"
 #include "spinlock.h"
