+++ kernel/proc/proc.c
@@ -1,7 +1,7 @@
 #include "types.h"
 #include "param.h"
 #include "memlayout.h"
-#include "riscv.h"
+#include "arch/riscv.h"
 #include "spinlock.h"
 #include "proc.h"
 #include "defs.h"
