+++ kernel/sys/sysnet.c
@@ -5,14 +5,14 @@
 #include "types.h"
 #include "param.h"
 #include "memlayout.h"
-#include "riscv.h"
+#include "arch/riscv.h"
 #include "spinlock.h"
 #include "proc.h"
 #include "defs.h"
 #include "fs.h"
 #include "sleeplock.h"
 #include "file.h"
-#include "net.h"
+#include "net/mbuf.h"
 
 struct sock {
   struct sock *next; // the next socket in the list
